fix: cli cmd
