+++ cmd/awscli/main.go
@@ -15,140 +15,9 @@
 package main
 
 import (
-	"fmt"
-	"net/http"
-	"net/url"
-	"os"
-
-	"golang.org/x/net/http/httpproxy"
-
-	"yunion.io/x/pkg/errors"
-	"yunion.io/x/pkg/util/shellutils"
-	"yunion.io/x/structarg"
-
-	"yunion.io/x/cloudmux/pkg/cloudprovider"
-	"yunion.io/x/cloudmux/pkg/multicloud/aws"
-	_ "yunion.io/x/cloudmux/pkg/multicloud/aws/shell"
+	"yunion.io/x/cloudmux/pkg/multicloud/aws/awscli"
 )
 
-type BaseOptions struct {
-	Debug      bool   `help:"debug mode"`
-	AccessUrl  string `help:"Access key" default:"$AWS_ACCESS_URL" choices:"ChinaCloud|InternationalCloud" metavar:"AWS_ACCESS_URL"`
-	AccessKey  string `help:"Access key" default:"$AWS_ACCESS_KEY" metavar:"AWS_ACCESS_KEY"`
-	Secret     string `help:"Secret" default:"$AWS_SECRET" metavar:"AWS_SECRET"`
-	RegionId   string `help:"RegionId" default:"$AWS_REGION" metavar:"AWS_REGION"`
-	AccountId  string `help:"Subaccount ID" default:"$AWS_ACCOUNT_ID" metavar:"AWS_ACCOUNT_ID"`
-	SUBCOMMAND string `help:"awscli subcommand" subcommand:"true"`
-}
-
-func getSubcommandParser() (*structarg.ArgumentParser, error) {
-	parse, e := structarg.NewArgumentParserWithHelp(&BaseOptions{},
-		"awscli",
-		"Command-line interface to aws API.",
-		`See "awscli COMMAND --help" for help on a specific command.`)
-
-	if e != nil {
-		return nil, e
-	}
-
-	subcmd := parse.GetSubcommand()
-	if subcmd == nil {
-		return nil, fmt.Errorf("No subcommand argument.")
-	}
-	for _, v := range shellutils.CommandTable {
-		_, e := subcmd.AddSubParserWithHelp(v.Options, v.Command, v.Desc, v.Callback)
-		if e != nil {
-			return nil, e
-		}
-	}
-	return parse, nil
-}
-
-func showErrorAndExit(e error) {
-	fmt.Fprintf(os.Stderr, "%s", e)
-	fmt.Fprintln(os.Stderr)
-	os.Exit(1)
-}
-
-func newClient(options *BaseOptions) (*aws.SRegion, error) {
-	if len(options.AccessKey) == 0 {
-		return nil, fmt.Errorf("Missing accessKey")
-	}
-
-	if len(options.Secret) == 0 {
-		return nil, fmt.Errorf("Missing secret")
-	}
-
-	cfg := &httpproxy.Config{
-		HTTPProxy:  os.Getenv("HTTP_PROXY"),
-		HTTPSProxy: os.Getenv("HTTPS_PROXY"),
-		NoProxy:    os.Getenv("NO_PROXY"),
-	}
-	cfgProxyFunc := cfg.ProxyFunc()
-	proxyFunc := func(req *http.Request) (*url.URL, error) {
-		return cfgProxyFunc(req.URL)
-	}
-
-	cli, err := aws.NewAwsClient(
-		aws.NewAwsClientConfig(
-			options.AccessUrl,
-			options.AccessKey,
-			options.Secret,
-			options.AccountId,
-		).Debug(options.Debug).
-			CloudproviderConfig(
-				cloudprovider.ProviderConfig{
-					ProxyFunc: proxyFunc,
-				},
-			),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	region, err := cli.GetRegion(options.RegionId)
-	if err != nil {
-		return nil, errors.Wrapf(err, "GetRegion(%s)", options.RegionId)
-	}
-
-	return region, nil
-}
-
 func main() {
-	parser, e := getSubcommandParser()
-	if e != nil {
-		showErrorAndExit(e)
-	}
-	e = parser.ParseArgs(os.Args[1:], false)
-	options := parser.Options().(*BaseOptions)
-
-	if parser.IsHelpSet() {
-		fmt.Print(parser.HelpString())
-		return
-	}
-	subcmd := parser.GetSubcommand()
-	subparser := subcmd.GetSubParser()
-	if e != nil || subparser == nil {
-		if subparser != nil {
-			fmt.Print(subparser.Usage())
-		} else {
-			fmt.Print(parser.Usage())
-		}
-		showErrorAndExit(e)
-		return
-	}
-	suboptions := subparser.Options()
-	if subparser.IsHelpSet() {
-		fmt.Print(subparser.HelpString())
-		return
-	}
-	var region *aws.SRegion
-	region, e = newClient(options)
-	if e != nil {
-		showErrorAndExit(e)
-	}
-	e = subcmd.Invoke(region, suboptions)
-	if e != nil {
-		showErrorAndExit(e)
-	}
+	awscli.Main()
 }
