+++ cmd/hcsocli/main.go
@@ -15,140 +15,9 @@
 package main
 
 import (
-	"fmt"
-	"net/http"
-	"net/url"
-	"os"
-
-	"golang.org/x/net/http/httpproxy"
-
-	"yunion.io/x/pkg/util/shellutils"
-	"yunion.io/x/structarg"
-
-	"yunion.io/x/cloudmux/pkg/cloudprovider"
-	huawei "yunion.io/x/cloudmux/pkg/multicloud/hcso"
-	_ "yunion.io/x/cloudmux/pkg/multicloud/hcso/shell"
+	"yunion.io/x/cloudmux/pkg/multicloud/hcso/hcsocli"
 )
 
-type BaseOptions struct {
-	cloudprovider.SHCSOEndpoints
-	Debug          bool   `help:"Show debug" default:"false"`
-	AccessKey      string `help:"Access key" default:"$HUAWEI_ACCESS_KEY" metavar:"HUAWEI_ACCESS_KEY"`
-	Secret         string `help:"Secret" default:"$HUAWEI_SECRET" metavar:"HUAWEI_SECRET"`
-	RegionId       string `help:"RegionId" default:"$HUAWEI_REGION" metavar:"HUAWEI_REGION"`
-	DEFAULT_REGION string `help:"Default Region" default:"$HUAWEI_DEFAULT_REGION" metavar:"HUAWEI_DEFAULT_REGION"`
-	ProjectId      string `help:"ProjectId" default:"$HUAWEI_PROJECT" metavar:"HUAWEI_PROJECT"`
-	SUBCOMMAND     string `help:"huaweicli subcommand" subcommand:"true"`
-}
-
-func getSubcommandParser() (*structarg.ArgumentParser, error) {
-	parse, e := structarg.NewArgumentParserWithHelp(&BaseOptions{},
-		"hcsocli",
-		"Command-line interface to huawei API.",
-		`See "hcsocli COMMAND --help" for help on a specific command.`)
-
-	if e != nil {
-		return nil, e
-	}
-
-	subcmd := parse.GetSubcommand()
-	if subcmd == nil {
-		return nil, fmt.Errorf("No subcommand argument.")
-	}
-	for _, v := range shellutils.CommandTable {
-		_, e := subcmd.AddSubParserWithHelp(v.Options, v.Command, v.Desc, v.Callback)
-		if e != nil {
-			return nil, e
-		}
-	}
-	return parse, nil
-}
-
-func showErrorAndExit(e error) {
-	fmt.Fprintf(os.Stderr, "%s", e)
-	fmt.Fprintln(os.Stderr)
-	os.Exit(1)
-}
-
-func newClient(options *BaseOptions) (*huawei.SRegion, error) {
-	if len(options.AccessKey) == 0 {
-		return nil, fmt.Errorf("Missing accessKey")
-	}
-
-	if len(options.Secret) == 0 {
-		return nil, fmt.Errorf("Missing secret")
-	}
-
-	cfg := &httpproxy.Config{
-		HTTPProxy:  os.Getenv("HTTP_PROXY"),
-		HTTPSProxy: os.Getenv("HTTPS_PROXY"),
-		NoProxy:    os.Getenv("NO_PROXY"),
-	}
-	cfgProxyFunc := cfg.ProxyFunc()
-	proxyFunc := func(req *http.Request) (*url.URL, error) {
-		return cfgProxyFunc(req.URL)
-	}
-
-	cli, err := huawei.NewHuaweiClient(
-		huawei.NewHuaweiClientConfig(
-			options.AccessKey,
-			options.Secret,
-			options.ProjectId,
-			&options.SHCSOEndpoints,
-		).Debug(options.Debug).
-			CloudproviderConfig(
-				cloudprovider.ProviderConfig{
-					ProxyFunc:     proxyFunc,
-					DefaultRegion: options.DEFAULT_REGION,
-				},
-			),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	region := cli.GetRegion(options.RegionId)
-	if region == nil {
-		return nil, fmt.Errorf("No such region %s", options.RegionId)
-	}
-
-	return region, nil
-}
-
 func main() {
-	parser, e := getSubcommandParser()
-	if e != nil {
-		showErrorAndExit(e)
-	}
-	e = parser.ParseArgs(os.Args[1:], false)
-	options := parser.Options().(*BaseOptions)
-
-	if parser.IsHelpSet() {
-		fmt.Print(parser.HelpString())
-		return
-	}
-	subcmd := parser.GetSubcommand()
-	subparser := subcmd.GetSubParser()
-	if e != nil || subparser == nil {
-		if subparser != nil {
-			fmt.Print(subparser.Usage())
-		} else {
-			fmt.Print(parser.Usage())
-		}
-		showErrorAndExit(e)
-	}
-	suboptions := subparser.Options()
-	if subparser.IsHelpSet() {
-		fmt.Print(subparser.HelpString())
-		return
-	}
-	var region *huawei.SRegion
-	region, e = newClient(options)
-	if e != nil {
-		showErrorAndExit(e)
-	}
-	e = subcmd.Invoke(region, suboptions)
-	if e != nil {
-		showErrorAndExit(e)
-	}
+	hcsocli.Main()
 }
