+++ pkg/multicloud/cmx/shell/instance.go
@@ -1,0 +1,33 @@
+// Copyright 2019 Yunion
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package shell
+
+import (
+	"yunion.io/x/cloudmux/pkg/cloudprovider"
+)
+
+func init() {
+	cmd := NewCommand("instance")
+
+	type InstanceListOptions struct {
+		ListBaseOptions
+		HostBaseOptions
+
+		Id []string `help:"IDs of instances to show"`
+	}
+	HostR[InstanceListOptions](cmd).List("list", "List instances", func(cli cloudprovider.ICloudHost, _ *InstanceListOptions) (any, error) {
+		return cli.GetIVMs()
+	})
+}
