+++ pkg/multicloud/ecloud/ecloudcli/main.go
@@ -1,0 +1,147 @@
+// Copyright 2019 Yunion
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package ecloudcli
+
+import (
+	"fmt"
+	"net/http"
+	"net/url"
+	"os"
+
+	"golang.org/x/net/http/httpproxy"
+
+	"yunion.io/x/pkg/util/shellutils"
+	"yunion.io/x/structarg"
+
+	"yunion.io/x/cloudmux/pkg/cloudprovider"
+	"yunion.io/x/cloudmux/pkg/multicloud/ecloud"
+	_ "yunion.io/x/cloudmux/pkg/multicloud/ecloud/shell"
+)
+
+type Options struct {
+	Debug        bool   `help:"Show debug" default:"false"`
+	AccessKey    string `help:"Access key" default:"$ECLOUD_ACCESS_KEY"`
+	AccessSecret string `help:"Secret" default:"$ECLOUD_ACCESS_SECRET"`
+	RegionId     string `help:"RegionId" default:"$ECLOUD_REGION"`
+	SUBCOMMAND   string `help:"ecloudcli subcommand" subcommand:"true"`
+}
+
+func getSubcommandParser() (*structarg.ArgumentParser, error) {
+	parse, e := structarg.NewArgumentParserWithHelp(&Options{},
+		"ecloudcli",
+		"Command-line interface to ecloud API.",
+		`See "ecloudcli COMMAND --help" for help on a specific command.`)
+
+	if e != nil {
+		return nil, e
+	}
+
+	subcmd := parse.GetSubcommand()
+	if subcmd == nil {
+		return nil, fmt.Errorf("No subcommand argument.")
+	}
+	for _, v := range shellutils.CommandTable {
+		_, e := subcmd.AddSubParserWithHelp(v.Options, v.Command, v.Desc, v.Callback)
+		if e != nil {
+			return nil, e
+		}
+	}
+	return parse, nil
+}
+
+func showErrorAndExit(e error) {
+	fmt.Fprintf(os.Stderr, "%s", e)
+	fmt.Fprintln(os.Stderr)
+	os.Exit(1)
+}
+
+func newClient(options *Options) (*ecloud.SRegion, error) {
+	if len(options.AccessKey) == 0 {
+		return nil, fmt.Errorf("Missing access key")
+	}
+
+	if len(options.AccessSecret) == 0 {
+		return nil, fmt.Errorf("Missing access secret")
+	}
+
+	cfg := &httpproxy.Config{
+		HTTPProxy:  os.Getenv("HTTP_PROXY"),
+		HTTPSProxy: os.Getenv("HTTPS_PROXY"),
+		NoProxy:    os.Getenv("NO_PROXY"),
+	}
+	cfgProxyFunc := cfg.ProxyFunc()
+	proxyFunc := func(req *http.Request) (*url.URL, error) {
+		return cfgProxyFunc(req.URL)
+	}
+
+	cli, err := ecloud.NewEcloudClient(
+		ecloud.NewEcloudClientConfig(
+			ecloud.NewRamRoleSigner(options.AccessKey, options.AccessSecret),
+		).SetDebug(options.Debug).
+			SetCloudproviderConfig(
+				cloudprovider.ProviderConfig{
+					ProxyFunc: proxyFunc,
+				},
+			),
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	region, err := cli.GetRegionById(options.RegionId)
+	if err != nil {
+		return nil, err
+	}
+
+	return region, nil
+}
+
+func Main() {
+	parser, e := getSubcommandParser()
+	if e != nil {
+		showErrorAndExit(e)
+	}
+	e = parser.ParseArgs(os.Args[1:], false)
+	options := parser.Options().(*Options)
+
+	if parser.IsHelpSet() {
+		fmt.Print(parser.HelpString())
+		return
+	}
+	subcmd := parser.GetSubcommand()
+	subparser := subcmd.GetSubParser()
+	if e != nil || subparser == nil {
+		if subparser != nil {
+			fmt.Print(subparser.Usage())
+		} else {
+			fmt.Print(parser.Usage())
+		}
+		showErrorAndExit(e)
+	}
+	suboptions := subparser.Options()
+	if subparser.IsHelpSet() {
+		fmt.Print(subparser.HelpString())
+		return
+	}
+	var region *ecloud.SRegion
+	region, e = newClient(options)
+	if e != nil {
+		showErrorAndExit(e)
+	}
+	e = subcmd.Invoke(region, suboptions)
+	if e != nil {
+		showErrorAndExit(e)
+	}
+}
